Linted

+++ src/robomaster.cpp
@@ -1,27 +1,34 @@
 #include <boost/asio.hpp>
 #include <boost/thread/thread.hpp>
-#include <spdlog/spdlog.h>
+
+#include "spdlog/spdlog.h"
+
 // #include "spdlog/cfg/env.h"
 
 #include "robomaster.hpp"
 
-// DONE(not urgent): make Robot decoupled from the protocol (so that in future I can use it without protocol/real time/...)
-// for that:
-// - [x] set a callback (or a callback list) to be called at the end of an update step (to be set in the RM constructor)
-// - [x] refactor actions [seems the most difficult], splitting the robot specific and the protocol specific part
-// (how to link them)
+// DONE(not urgent): make Robot decoupled from the protocol (so that in future I can use it without
+// protocol/real time/...) for that:
+// - [x] set a callback (or a callback list) to be called at the end of an update step (to be set in
+// the RM constructor)
+// - [x] refactor actions [seems the most difficult], splitting the robot specific and the protocol
+// specific part (how to link them)
 // - [ ] then expose (actions, commands, and state) in lua too (first actions and commands)
-// Should be able to compile robot.c in a library without the protocol (and maybe instantiate is in the plugin (remote_api=<bool>))
+// Should be able to compile robot.c in a library without the protocol (and maybe instantiate is in
+// the plugin (remote_api=<bool>))
 
-
-RoboMaster::RoboMaster(std::shared_ptr<boost::asio::io_context> _io_context, Robot * _robot, std::string serial_number,
-                       bool udp_video_stream, long video_stream_bitrate)
-: io_context(_io_context ? _io_context : std::make_shared<boost::asio::io_context>()),
-  robot(_robot), discovery(io_context.get(), serial_number), conn(io_context.get(), robot, 30030),
-  cmds(io_context.get(), robot, this, 20020) {
+RoboMaster::RoboMaster(std::shared_ptr<boost::asio::io_context> _io_context, Robot *_robot,
+                       std::string serial_number, bool udp_video_stream,
+                       unsigned video_stream_bitrate)
+    : io_context(_io_context ? _io_context : std::make_shared<boost::asio::io_context>())
+    , robot(_robot)
+    , discovery(io_context.get(), serial_number)
+    , conn(io_context.get(), robot, 30030)
+    , cmds(io_context.get(), robot, this, 20020) {
   spdlog::set_level(spdlog::level::info);
   spdlog::info("Creating a RobotMaster with serial number {}", serial_number);
-  video = VideoStreamer::create_video_streamer(io_context.get(), robot, udp_video_stream, video_stream_bitrate);
+  video = VideoStreamer::create_video_streamer(io_context.get(), robot, udp_video_stream,
+                                               video_stream_bitrate);
   // spdlog::cfg::load_env_levels();
   discovery.start();
 
@@ -31,16 +38,15 @@ RoboMaster::RoboMaster(std::shared_ptr<boost::asio::io_context> _io_context, Rob
 void RoboMaster::do_step(float time_step) {
   discovery.do_step(time_step);
   cmds.do_step(time_step);
-  if(video)
+  if (video)
     video->do_step(time_step);
 }
 
 void RoboMaster::spin(bool thread) {
-  if(thread) {
+  if (thread) {
     spdlog::info("Start IO spinning in thread");
     t = new boost::thread(boost::bind(&boost::asio::io_context::run, io_context));
-  }
-  else {
+  } else {
     io_context->run();
   }
 }
